cohorts/2023/week_6_stream_processing/kafka/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
pandas==1.5.3
orjson==3.8.7
tqdm==4.65.0